+++ client.go
@@ -90,6 +90,10 @@ type Client interface {
 
 	// GetRolePermissions gets permissions of a role
 	GetRolePermissions(roleID string, opts ...Option) (perms []Permission, err error)
+
+	// GetClientInformation gets IAM client information,
+	// it will look into cache first, if not found then fetch it to IAM.
+	GetClientInformation(namespace string, clientID string, opts ...Option) (*ClientInformation, error)
 }
 
 type Options struct {